cryptocloud/core.py
# from Crypto.PublicKey import ECC
# from phe import paillier
# import sqlite3
# import random

# class PairingGroupMock:
#     """Simplified mock of pairing group operations"""
#     def __init__(self, curve_name):
#         self.curve = curve_name
        
#     def random(self, group_type='G1'):
#         """Generate mock group elements"""
#         return random.randint(1, 2**32)  # Return a random number
    
#     def __call__(self, element):
#         """Mock group operation"""
#         return element

# class CryptCloudPlus:
#     def __init__(self):
#         self.group = PairingGroupMock('SS512-mock')
#         self.g = self.group.random('G1')  # Now passing group_type
#         self.alpha = self.group.random('G1')  # And here
#         self.pub_key, self.priv_key = paillier.generate_keypair(n_length=512)
#         self.RL = []
#         self._init_db()
    
#     # ... [rest of your methods remain the same] ...

#     def _init_db(self):
#         self.conn = sqlite3.connect(':memory:')
#         self.conn.execute('''CREATE TABLE audit_log 
#                             (id INTEGER PRIMARY KEY, commitment BLOB)''')

#     def keygen(self, user_id, attributes):
#         """Generate user key with traceable ID commitment"""
#         id_commitment = self.pub_key.encrypt(user_id)
#         self.conn.execute('INSERT INTO audit_log VALUES (?, ?)', 
#                          (user_id, str(id_commitment.ciphertext())))
        
#         # Simplified cryptographic operations
#         t = self.group.random()
#         K = (self.g * (self.alpha + t)) % (2**256)  # Mock operation
        
#         return {
#             'sk': K,
#             'commitment': id_commitment,
#             'attrs': attributes
#         }

#     def encrypt(self, message, policy):
#         """Encrypt data with access policy"""
#         s = self.group.random()
#         # Simplified encryption
#         C = (message + (self.g * (self.alpha * s))) % (2**256)
        
#         return {
#             'C': C,
#             'policy': policy,
#             'RL': self.RL
#         }

#     def decrypt(self, user_key, ciphertext):
#         """Attempt decryption with access checks"""
#         try:
#             user_id = self.priv_key.decrypt(user_key['commitment'])
#             if user_id in ciphertext['RL']:
#                 return "🔒 ACCESS DENIED (Revoked)"
#             if not set(user_key['attrs']).issuperset(ciphertext['policy']):
#                 return "🔒 ACCESS DENIED (Policy)"
#             return "🔓 ACCESS GRANTED: SecretData"
#         except:
#             return "🔒 DECRYPTION FAILED"

#     def trace(self, leaked_key):
#         """Trace key to user ID"""
#         return self.priv_key.decrypt(leaked_key['commitment'])

#     def revoke(self, user_id):
#         """Add user to revocation list"""
#         self.RL.append(user_id)

# # Example usage
# if __name__ == "__main__":
#     cc = CryptCloudPlus()
    
#     # Create users
#     faculty = cc.keygen(101, ['role=Faculty', 'dept=CS'])
#     student = cc.keygen(102, ['role=Student'])
    
#     # Test cases
#     policy = ['role=Faculty']
#     ciphertext = cc.encrypt(12345, policy)
    
#     print("Faculty access:", cc.decrypt(faculty, ciphertext))
#     print("Student access:", cc.decrypt(student, ciphertext))
    
#     # Revocation test
#     cc.revoke(101)
#     print("After revocation:", cc.decrypt(faculty, ciphertext))


from phe import paillier
import sqlite3
import random
from Crypto.Hash import SHA256

class CryptCloudPlus:
    def __init__(self):
        # Simplified crypto setup
        self.pub_key, self.priv_key = paillier.PaillierPublicKey, paillier.PaillierPrivateKey
        self.pub_key, self.priv_key = paillier.generate_paillier_keypair(n_length=512)
        self.RL = []  # Revocation list
        self._init_db()
        self.g = random.getrandbits(256)  # Mock generator
        self.alpha = random.getrandbits(256)  # Master secret

    def _init_db(self):
        self.conn = sqlite3.connect(':memory:')
        self.conn.execute('''CREATE TABLE audit_log 
                           (id INTEGER PRIMARY KEY, commitment BLOB)''')

    def keygen(self, user_id, attributes):
        """Generate user key with traceable ID"""
        id_commitment = self.pub_key.encrypt(user_id)
        self.conn.execute('INSERT INTO audit_log VALUES (?, ?)',
                         (user_id, str(id_commitment.ciphertext())))
        return {
            'sk': (self.g + self.alpha + random.getrandbits(256)) % (2**256),
            'commitment': id_commitment,
            'attrs': attributes
        }

    def encrypt(self, message, policy):
        """Simplified encryption"""
        return {
            'C': f"ENC_{message}",
            'policy': policy,
            'RL': self.RL
        }

    def decrypt(self, user_key, ciphertext):
        """Access control checks"""
        user_id = self.priv_key.decrypt(user_key['commitment'])
        if user_id in ciphertext['RL']:
            return "🔒 ACCESS DENIED (Revoked)"
        if not set(user_key['attrs']).issuperset(ciphertext['policy']):
            return "🔒 ACCESS DENIED (Policy)"
        return f"🔓 ACCESS GRANTED: {ciphertext['C'][4:]}"

    def trace(self, leaked_key):
        return self.priv_key.decrypt(leaked_key['commitment'])

    def revoke(self, user_id):
        self.RL.append(user_id)